refactor: userqueryservice 클래스분리 및 리펙토링
