+++ user-service/src/main/java/org/sixpang/userservice/application/service/UserQueryService.java
@@ -1,7 +1,9 @@
 package org.sixpang.userservice.application.service;
 
 import lombok.RequiredArgsConstructor;
-import org.sixpang.userservice.application.dto.UserQueryDto;
+import org.sixpang.userservice.application.dto.query.UserDetail;
+import org.sixpang.userservice.application.dto.query.UserInfo;
+import org.sixpang.userservice.application.dto.query.AuthUser;
 import org.sixpang.userservice.domain.model.entity.User;
 import org.sixpang.userservice.domain.repository.UserRepository;
 import org.sixpang.userservice.exception.UserErrorCode;
@@ -21,30 +23,30 @@ public class UserQueryService {
     private final UserRepository userRepository;
 
     /**단건 조회 (상세)**/
-    public UserQueryDto.UserDetail getUser(UUID userId) {
+    public UserDetail getUser(UUID userId) {
 
         User user = userRepository.findByIdAndDeletedAtIsNull(userId)
                 .orElseThrow(() -> new UserException(UserErrorCode.USER_NOT_FOUND));
 
         // 코드 리뷰: DTO 생성 책임을 DTO로 이동 (from 메서드 사용)
-        return UserQueryDto.UserDetail.from(user);
+        return UserDetail.from(user);
     }
 
     /**목록 조회**/
-    public Page<UserQueryDto.UserInfo> getUsers(Pageable pageable) {
+    public Page<UserInfo> getUsers(Pageable pageable) {
 
         return userRepository.findAllByDeletedAtIsNull(pageable)
                 // 코드 리뷰: 객체 생성 로직을 DTO로 위임
-                .map(UserQueryDto.UserInfo::from);
+                .map(UserInfo::from);
     }
 
     /** 로그인용 이메일 조회 (AuthService에서 사용) **/
-    public UserQueryDto.AuthUser getUserByEmail(String email) {
+    public AuthUser getUserByEmail(String email) {
 
         User user = userRepository.findByEmailAndDeletedAtIsNull(email)
                 .orElseThrow(() -> new UserException(UserErrorCode.USER_NOT_FOUND));
 
         // 코드 리뷰: DTO 생성 책임을 DTO로 이동
-        return UserQueryDto.AuthUser.from(user);
+        return AuthUser.from(user);
     }
 }
